pyproject.toml
[tool.poetry]
name = "hordeqt-py"
version = "0.1.0"
description = ""
authors = ["Unit1208 <unit1208@national.shitposting.agency>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
black = "^24.8.0"
autopep8 = "^2.3.1"
pyside6 = "^6.7.2"
keyring = "^25.3.0"
keyrings-alt = "^5.0.2"
requests = "^2.32.3"
nuitka = "^2.4.8"
zstandard = "^0.23.0"
human-readable = "^1.3.4"
pyqt-toast-notification = "^1.3.2"
pillow = "^10.4.0"
coloredlogs = "^15.0.1"
colorama = "^0.4.6"

[tool.poetry.scripts]

build = "hordeqt_py.build:main"
regen = "hordeqt_py.build:convert_uic_files"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nuitka]
standalone = true
onefile = true
remove-output = true
output-dir = "dist"

[tool.black]
extend-exclude='''
(
^/hordeqt_py/ui_[_a-zA-Z]+\.py
)'''